fix(320): solve architecture violation issues for circular dependencies
