+++ app/(main)/(authenticated)/(tabs)/home-page.tsx
@@ -1,4 +1,4 @@
-import { UpcomingTripPage } from '@/features/trips';
+import { UpcomingTripPage } from '@/features/trips/pages';
 
 export default function HomePageTab() {
   return <UpcomingTripPage />;
