+++ app/(main)/(authenticated)/(tabs)/profile.tsx
@@ -1,4 +1,4 @@
-import { ProfilePage } from '@/features/profile';
+import { ProfilePage } from '@/features/profile/pages';
 
 export default function Profile() {
   return <ProfilePage />;
