+++ app/(main)/(authenticated)/create-trip/activity-details.tsx
@@ -1,4 +1,4 @@
-import { ActivityDetailsPage } from '@/features/trips';
+import { ActivityDetailsPage } from '@/features/trips/pages';
 
 const ActivityDetails = () => {
   return <ActivityDetailsPage />;
