+++ app/(main)/(authenticated)/create-trip/generate-trip.tsx
@@ -1,4 +1,4 @@
-import { GenerateTripPage } from '@/features/trip-generation';
+import { GenerateTripPage } from '@/features/trip-generation/pages';
 
 const GenerateTrip = () => {
   return <GenerateTripPage />;
