+++ app/(main)/(authenticated)/create-trip/review-trip.tsx
@@ -1,4 +1,4 @@
-import { ReviewTripPage } from '@/features/trip-generation';
+import { ReviewTripPage } from '@/features/trip-generation/pages';
 
 const ReviewTrip = () => {
   return <ReviewTripPage />;
