+++ app/(main)/(authenticated)/create-trip/select-dates.tsx
@@ -1,4 +1,4 @@
-import { SelectDatesPage } from '@/features/trip-generation';
+import { SelectDatesPage } from '@/features/trip-generation/pages';
 
 const SelectDates = () => {
   return <SelectDatesPage />;
