+++ app/(main)/(authenticated)/create-trip/select-traveler.tsx
@@ -1,4 +1,4 @@
-import { SelectTravelersPage } from '@/features/trip-generation';
+import { SelectTravelersPage } from '@/features/trip-generation/pages';
 
 const SelectTraveler = () => {
   return <SelectTravelersPage />;
