+++ app/(main)/(authenticated)/create-trip/trip-details.tsx
@@ -1,4 +1,4 @@
-import { TripDetailsPage } from '@/features/trips';
+import { TripDetailsPage } from '@/features/trips/pages';
 
 const TripDetails = () => {
   return <TripDetailsPage />;
