+++ app/(main)/(authenticated)/home-page/show-all-trips.tsx
@@ -1,4 +1,4 @@
-import { TripListPage } from '@/features/trips';
+import { TripListPage } from '@/features/trips/pages';
 
 export default function ShowAllTrips() {
   return <TripListPage />;
