+++ app/(main)/(authenticated)/profile/change-language.tsx
@@ -1,4 +1,4 @@
-import { ChangeLanguagePage } from '@/features/profile';
+import { ChangeLanguagePage } from '@/features/profile/pages';
 
 export default function ChangeLanguage() {
   return <ChangeLanguagePage />;
