+++ app/(main)/(login)/sign-in.tsx
@@ -1,4 +1,4 @@
-import { SignInPage } from '@/features/auth';
+import { SignInPage } from '@/features/auth/pages';
 
 export default function SignIn() {
   return <SignInPage />;
