+++ app/(main)/(login)/sign-up.tsx
@@ -1,4 +1,4 @@
-import { SignUpPage } from '@/features/auth';
+import { SignUpPage } from '@/features/auth/pages';
 
 export default function SignUp() {
   return <SignUpPage />;
