+++ app/(main)/(login)/welcome.tsx
@@ -1,4 +1,4 @@
-import { WelcomePage } from '@/features/auth';
+import { WelcomePage } from '@/features/auth/pages';
 
 export default function Welcome() {
   return <WelcomePage />;
