+++ features/auth/index.ts
@@ -2,6 +2,3 @@ import type { FeatureTier } from '@/features/core/featureTier';
 export const FEATURE_TIER: FeatureTier = 1;
 
 export { useResetPassword } from '@/features/auth/facades/useResetPassword';
-export { SignInPage } from '@/features/auth/ui/pages/SignInPage/SignInPage';
-export { SignUpPage } from '@/features/auth/ui/pages/SignUpPage/SignUpPage';
-export { WelcomePage } from '@/features/auth/ui/pages/WelcomePage/WelcomePage';
