+++ features/auth/pages.ts
@@ -1,0 +1,3 @@
+export { SignInPage } from '@/features/auth/ui/pages/SignInPage/SignInPage';
+export { SignUpPage } from '@/features/auth/ui/pages/SignUpPage/SignUpPage';
+export { WelcomePage } from '@/features/auth/ui/pages/WelcomePage/WelcomePage';
