+++ features/profile/index.ts
@@ -3,5 +3,3 @@ export const FEATURE_TIER: FeatureTier = 3;
 
 export { useDeleteAccount } from '@/features/profile/facades/useDeleteAccount';
 export { useProfileData } from '@/features/profile/facades/useProfileData';
-export { ChangeLanguagePage } from '@/features/profile/ui/pages/ChangeLanguagePage/ChangeLanguagePage';
-export { ProfilePage } from '@/features/profile/ui/pages/ProfilePage/ProfilePage';
