+++ features/profile/pages.ts
@@ -1,0 +1,2 @@
+export { ChangeLanguagePage } from '@/features/profile/ui/pages/ChangeLanguagePage/ChangeLanguagePage';
+export { ProfilePage } from '@/features/profile/ui/pages/ProfilePage/ProfilePage';
