+++ features/trip-generation/pages.ts
@@ -1,0 +1,6 @@
+export { GenerateTripPage } from '@/features/trip-generation/ui/pages/GenerateTripPage/GenerateTripPage';
+export { ReviewTripPage } from '@/features/trip-generation/ui/pages/ReviewTripPage/ReviewTripPage';
+export { SearchPlacePage } from '@/features/trip-generation/ui/pages/SearchPlacePage/SearchPlacePage';
+export { SelectBudgetPage } from '@/features/trip-generation/ui/pages/SelectBudgetPage/SelectBudgetPage';
+export { SelectDatesPage } from '@/features/trip-generation/ui/pages/SelectDatesPage/SelectDatesPage';
+export { SelectTravelersPage } from '@/features/trip-generation/ui/pages/SelectTravelersPage/SelectTravelersPage';
