+++ features/trips/index.ts
@@ -12,7 +12,3 @@ export { useAddTrip } from '@/features/trips/facades/useAddTrip';
 export { useDeleteAllTrips } from '@/features/trips/facades/useDeleteAllTrips';
 export { useGetFavoriteTrips } from '@/features/trips/facades/useGetFavoriteTrips';
 export { useGetTrips } from '@/features/trips/facades/useGetTrips';
-export { TripListPage } from '@/features/trips/ui/pages/TripListPage/TripListPage';
-export { TripDetailsPage } from '@/features/trips/ui/pages/TripDetailsPage/TripDetailsPage';
-export { ActivityDetailsPage } from '@/features/trips/ui/pages/ActivityDetailsPage/ActivityDetailsPage';
-export { UpcomingTripPage } from '@/features/trips/ui/pages/UpcomingTripPage/UpcomingTripPage';
