+++ features/trips/pages.ts
@@ -1,0 +1,4 @@
+export { ActivityDetailsPage } from '@/features/trips/ui/pages/ActivityDetailsPage/ActivityDetailsPage';
+export { TripDetailsPage } from '@/features/trips/ui/pages/TripDetailsPage/TripDetailsPage';
+export { TripListPage } from '@/features/trips/ui/pages/TripListPage/TripListPage';
+export { UpcomingTripPage } from '@/features/trips/ui/pages/UpcomingTripPage/UpcomingTripPage';
