+++ ui/components/basic/CustomCard/CustomCard.style.ts
@@ -1,8 +1,13 @@
-import { StyleSheet } from 'react-native';
+import { StyleSheet, type ViewStyle } from 'react-native';
 
-import type { CustomCardStyle } from '@/ui/components/basic/CustomCard/CustomCard.logic';
 import { spacing } from '@/ui/style/dimensions/spacing';
 import { opacity } from '@/ui/style/opacity';
+
+export type CustomCardStyle = {
+  backgroundColor: ViewStyle['backgroundColor'];
+  borderColor: ViewStyle['borderColor'];
+  borderWidth: ViewStyle['borderWidth'];
+};
 
 export const styles = ({ backgroundColor, borderColor, borderWidth }: CustomCardStyle) =>
   StyleSheet.create({
